+++ modules/loki/main.tf
@@ -61,9 +61,9 @@ resource "helm_release" "loki" {
   namespace  = var.loki.namespace
 
   create_namespace = false
-  wait             = true
-  wait_for_jobs    = true
-  timeout          = 600
+  wait             = var.loki.wait
+  wait_for_jobs    = var.loki.wait_for_jobs
+  timeout          = var.loki.timeout
 
   values = [
     templatefile("${path.module}/helm-values/loki.yaml.tftpl", {